User authentication (jwt) BE & FE

+++ controlers/expenseControlls.js
@@ -3,21 +3,19 @@ const path = require('path');
 const rootDir = require('../util/path');
 
 exports.createExpense = async (req,res,next) => {
+    const { expenseAmt, description, category } = req.body;
         try {
-            const expenseAmt = req.body.expenseAmt;
-            const description = req.body.description;
-            const category = req.body.category;
-
             if(expenseAmt == '' || description == '' || category == ''){
                 res.json({ error: 'Please fill details'});
             }
     
             const expense = await Expense.create({
-               expenseAmt: expenseAmt,
-               description: description,
-               category: category
+               expenseAmt,
+               description,
+               category,
+               userId: req.user.id
             });
-            res.status(200).redirect('/insert-expense');
+            res.status(200).json({expenseCreated: expense});
             // res.json({expenseCreated: expense});
             
         } catch (error) { console.log(error) }
@@ -26,24 +24,23 @@ exports.createExpense = async (req,res,next) => {
 
 exports.getExpenses = async (req,res,next) => {
     try {
-        const expenses = await Expense.findAll();
+        const expenses = await Expense.findAll({where: {userId: req.user.id}});
         res.json(expenses);
     } catch (error) {
         console.log(err);
     }
 }
 
-exports.getExpensePage = (req,res,next) => {
-    res.sendfile(path.join(rootDir, 'public','expense.html'))
-}
-
 
 exports.deleteExpense = async (req,res,next) => {
     try {
         const expenseId = req.params.expenseId;
-        await Expense.destroy({where: {id: expenseId}});
+        const noOfrecords = await Expense.destroy({where: {id: expenseId,userId: req.user.id}});
+        if(noOfrecords == 0){
+            res.status(404).json({success: false, msg: 'You cant delete other people expense'})
+        }
         // res.redirect('/');
-        res.json({message:"expense deleted successfully"});
+        return res.status(200).json({message:"expense deleted successfully"});
     } catch (error) {
         console.log(error);
     }
